Issue#8 (#96)
* issue #8

CD pipeline for issue #8

* Update cd-infra-common.yml

Update to CD file for issue #8

* Issue #8 update

updating Trigger to None/Maunal trigger/PR. This will change down the road, but a fix for now.

* Issue #8 Removed SubID and Validation

Issue #8 - Updated cd-infra-exthub/cd-infra-common. I removed the template validation, and also removed the subscription ID as we are using the service connection.