YAML corrected and docs updated
We had some mismatch of using echo and extensionecho.
But the big problem is that the image: $ACR_NAME is attempting to be expanded on the cluster (never going to work).  We have to do that before we apply the template.

* Naming samples with sample-name

I think it should be a best practice to name them with sample-name for easy identification in the kubectl logs.

* Name change

Changed the name to sample-echo.yaml to match the naming.

* Updated docs to tell user to replace {{ACR_NAME}} with their actual registry

Kubectl cannot do variable expansion.  We may fix this later using Helm.